+++ CHANGES.md
@@ -3,4 +3,5 @@ Gwtlib Changes
 
 master (top: more recent changes)
 ----------------------------------
+- add SimpleMessageDialogBoxPresenter and MessageWithCommandDialogBoxPresenter
 - add CustomHTMLPanel and FireAndForgetCallback